Update-555d3bee-38cb-42da-8eae-f34cb24557c1

+++ src/main/java/com/bss/inc/cameraon/utils/net/angryscan/feeders/RangeFeeder.java
@@ -63,8 +63,8 @@ public class RangeFeeder extends AbstractFeeder {
         return new ScanningSubject(prevIP);
     }
 
-    public float percentageComplete() {
-        return (float)percentageComplete;
+    public double percentageComplete() {
+        return percentageComplete;
     }
 
     public String getInfo() {
